[fix] Fixed basic locomotive scroll behavior by moving the dynamic component and changing the target container.
